Created alternate activation method by just clicking the extension icon
Wrote the code that could be used to do the document > markdown > download
chain directly upon the user clicking the extension icon (as opposed to
opening a popup with a button).
Left it commented-out, though, because the user experience feels less
smooth this way (less clear feedback, easier to accidentally do a thing).